+++ Android/EasyBudget/app/src/main/res/layout/activity_main.xml
@@ -13,7 +13,6 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 -->
-
 <FrameLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
@@ -40,176 +39,12 @@
 
         </com.google.android.material.appbar.AppBarLayout>
 
-        <FrameLayout
-            android:id="@+id/calendarView"
+        <androidx.fragment.app.FragmentContainerView
+            android:id="@+id/mainFragmentContainer"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"/>
-
-        <LinearLayout
-            android:id="@+id/budgetLineContainer"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal"
-            android:gravity="center_horizontal"
-            android:paddingLeft="15dp"
-            android:paddingRight="15dp"
-            android:paddingTop="4dp"
-            android:paddingBottom="4dp"
-            android:layout_marginTop="4dp"
-            android:background="@color/budget_line_background_color">
-
-            <TextView android:id="@+id/budgetLine"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:textColor="@color/primary_text"
-                android:textSize="14sp"/>
-
-            <TextView android:id="@+id/budgetLineAmount"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:textSize="14sp"
-                android:textStyle="bold"
-                android:layout_marginStart="4dp"/>
-
-        </LinearLayout>
-
-        <LinearLayout
-            android:id="@+id/emptyExpensesRecyclerViewPlaceholder"
-            android:layout_width="match_parent"
-            android:layout_height="0dip"
-            android:layout_weight="1"
-            android:orientation="vertical"
-            android:gravity="center_horizontal"
-            android:visibility="gone"
-            android:paddingTop="15dp">
-
-            <ImageView
-                android:layout_width="48dp"
-                android:layout_height="48dp"
-                android:src="@drawable/ic_wallet"
-                android:alpha="0.6"
-                android:layout_marginBottom="5dp"/>
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:textColor="@color/secondary_text"
-                android:textSize="14sp"
-                android:text="@string/no_expense_for_today" />
-
-        </LinearLayout>
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/expensesRecyclerView"
-            android:scrollbars="vertical"
-            android:layout_width="match_parent"
-            android:layout_height="0dip"
-            android:layout_weight="1"
-            android:background="@color/window_background"
-            android:clipToPadding="false"
-            android:paddingBottom="100dp"
-            android:scrollbarStyle="outsideOverlay"/>
+            android:layout_height="match_parent" />
 
     </LinearLayout>
-
-    <View
-        android:id="@+id/fab_choices_background"
-        android:visibility="gone"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/menu_background_overlay_color"/>
-
-    <androidx.coordinatorlayout.widget.CoordinatorLayout
-        android:id="@+id/coordinatorLayout"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:clipChildren="false">
-
-        <LinearLayout
-            android:id="@+id/fab_choices_linear_layout"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            android:layout_marginBottom="20dp"
-            android:layout_marginRight="20dp"
-            android:layout_gravity="right|bottom"
-            android:gravity="right"
-            android:clipChildren="false"
-            android:clipToPadding="false"
-            app:layout_behavior="com.benoitletondor.easybudgetapp.view.main.FloatingActionButtonBehavior">
-
-            <LinearLayout
-                android:id="@+id/fab_new_recurring_expense_container"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:orientation="horizontal"
-                android:gravity="center_vertical"
-                android:layout_marginBottom="10dp"
-                android:clipChildren="false"
-                android:clipToPadding="false"
-                android:visibility="gone">
-
-                <TextView
-                    android:id="@+id/fab_new_recurring_expense_text"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="@string/fab_add_monthly_expense"
-                    android:textSize="15sp"
-                    android:layout_marginRight="10dp"
-                    style="@style/fab_labels_style"/>
-
-                <com.google.android.material.floatingactionbutton.FloatingActionButton
-                    android:id="@+id/fab_new_recurring_expense"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    app:backgroundTint="@color/fab_add_expense"
-                    app:srcCompat="@drawable/ic_autorenew_white"
-                    app:tint="@color/white"
-                    app:fabSize="mini" />
-
-            </LinearLayout>
-
-            <LinearLayout
-                android:id="@+id/fab_new_expense_container"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:orientation="horizontal"
-                android:gravity="center_vertical"
-                android:layout_marginBottom="20dp"
-                android:clipChildren="false"
-                android:clipToPadding="false"
-                android:visibility="gone">
-
-                <TextView
-                    android:id="@+id/fab_new_expense_text"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="@string/fab_add_expense"
-                    android:textSize="15sp"
-                    android:layout_marginRight="10dp"
-                    style="@style/fab_labels_style"/>
-
-                <com.google.android.material.floatingactionbutton.FloatingActionButton
-                    android:id="@+id/fab_new_expense"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    app:backgroundTint="@color/fab_add_monthly_expense"
-                    app:srcCompat="@drawable/ic_baseline_add_24"
-                    app:tint="@color/white"
-                    app:fabSize="mini" />
-
-            </LinearLayout>
-
-            <com.google.android.material.floatingactionbutton.FloatingActionButton
-                android:id="@+id/fab_choices"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                app:srcCompat="@drawable/ic_baseline_add_24"
-                app:backgroundTint="@color/home_fab_button_color"
-                app:tint="@color/white" />
-        </LinearLayout>
-
-    </androidx.coordinatorlayout.widget.CoordinatorLayout>
 
     <LinearLayout
         android:id="@+id/monthly_report_hint"
