+++ Android/EasyBudget/app/src/main/res/menu/menu_account.xml
@@ -20,22 +20,23 @@
     tools:context=".view.main.MainActivity">
 
     <item
-        android:id="@+id/action_become_premium"
-        android:title="@string/action_become_premium"
-        android:icon="@drawable/ic_baseline_star_24"
+        android:id="@+id/action_go_to_current_month"
+        android:title="@string/action_go_to_current_month"
+        android:icon="@drawable/ic_calendar_today"
+        android:orderInCategory="50"
         app:showAsAction="always" />
 
     <item
-        android:id="@+id/action_monthly_report"
-        android:title="@string/monthly_report_button_title"
-        android:icon="@drawable/ic_list_alt_24"
-        android:orderInCategory="60"
-        app:showAsAction="ifRoom" />
+        android:id="@+id/action_balance"
+        android:title="@string/action_balance"
+        android:orderInCategory="70"
+        app:showAsAction="never" />
 
     <item
-        android:id="@+id/action_settings"
-        android:title="@string/action_settings"
-        android:orderInCategory="100"
+        android:id="@+id/action_check_all_past_entries"
+        android:title="@string/action_mark_all_past_entries_as_checked"
+        android:icon="@drawable/ic_check_24dp"
+        android:orderInCategory="80"
         app:showAsAction="never" />
 
 </menu>
